+++ src/listeners/index.ts
@@ -1,4 +1,4 @@
 import "./androidSafeArea";
 import "./keyboardPageResizer";
 import "./statusTap";
-import "./network";
+import "./network/listener";
