+++ src/listeners/network/networkSlice.ts
